Move FSST symbol table logic to dedicated module
Refactor `crates/nest-format/src/encoding/fsst.rs` by extracting the
symbol table definition and logic into `fsst_table.rs`. Additionally,
remove debug timing instrumentation from `text_codec.rs`.
