+++ crates/nest-format/src/encoding/fsst.rs
@@ -17,6 +17,7 @@
 //! as surfaced by duckdb's research (255-entry table, 1-8 byte symbols ->
 //! 1-byte codes, 0xFF escape). NO code is vendored.
 
+use super::fsst_table::{SymbolTable, parse_table, serialize_table};
 use super::intpack::{IntpackReader, pack_u64s};
 use super::txt_streams::{build_canonical, malformed, write_container};
 
@@ -25,218 +26,6 @@ pub const TXT_STREAMS_V3: u8 = 2;
 
 /// escape code: the next byte in the code stream is emitted raw.
 const ESCAPE: u8 = 0xFF;
-/// max symbol length in bytes (codes 0..=254 may map to 1..=8 raw bytes).
-const MAX_SYMBOL_LEN: usize = 8;
-/// number of assignable codes (0..=254); 255 is the escape.
-const N_CODES: usize = 255;
-
-/// a built symbol table: `symbols[code]` is the byte string code `code`
-/// expands to. at most 255 entries. round-trips through [`serialize_table`].
-struct SymbolTable {
-    symbols: Vec<Vec<u8>>,
-    /// trie over the symbols for O(1)-per-byte longest-match encoding.
-    /// each node maps a next byte to (child index, code if this node is a
-    /// terminal symbol). the trie is built from the same deterministic
-    /// `symbols` vector, so the encode output is byte-identical.
-    trie: Trie,
-}
-
-/// a compact byte trie for greedy longest-match encoding. `next[byte]` gives
-/// an optional child index, and `code` records the symbol code (if any) at
-/// this node. the trie is rebuilt from the deterministic symbol table, so it
-/// does not affect byte-identicality of the encoded output.
-#[derive(Default)]
-struct Trie {
-    next: Vec<[Option<u16>; 256]>,
-    code: Vec<Option<u8>>,
-}
-
-impl Trie {
-    fn new() -> Self {
-        Self {
-            next: vec![[None; 256]],
-            code: vec![None],
-        }
-    }
-
-    fn insert(&mut self, sym: &[u8], code: u8) {
-        let mut node = 0usize;
-        for &b in sym {
-            let b = b as usize;
-            let child = self.next[node][b];
-            let child = child.unwrap_or_else(|| {
-                let idx = self.next.len() as u16;
-                self.next.push([None; 256]);
-                self.code.push(None);
-                self.next[node][b] = Some(idx);
-                idx
-            });
-            node = child as usize;
-        }
-        self.code[node] = Some(code);
-    }
-
-    /// longest symbol match starting at `input`. returns `(code, length)`.
-    fn longest_match(&self, input: &[u8]) -> Option<(u8, usize)> {
-        let mut node = 0usize;
-        let mut best: Option<(u8, usize)> = None;
-        for (i, &b) in input.iter().enumerate().take(MAX_SYMBOL_LEN) {
-            let child = self.next[node][b as usize]?;
-            node = child as usize;
-            if let Some(code) = self.code[node] {
-                best = Some((code, i + 1));
-            }
-        }
-        best
-    }
-}
-
-/// frequency trie for counting every 1..=MAX_SYMBOL_LEN substring without
-/// allocating per-substring keys. each node records the number of times the
-/// byte sequence that ends at it was seen as a full substring.
-#[derive(Default)]
-struct FreqTrie {
-    next: Vec<[Option<u32>; 256]>,
-    count: Vec<u64>,
-}
-
-impl FreqTrie {
-    fn new() -> Self {
-        Self {
-            next: vec![[None; 256]],
-            count: vec![0],
-        }
-    }
-
-    fn count_substrings(&mut self, corpus: &[u8], max_len: usize) {
-        let n = corpus.len();
-        for start in 0..n {
-            let max = (start + max_len).min(n);
-            let mut node = 0usize;
-            for end in start..max {
-                let b = corpus[end] as usize;
-                let child = self.next[node][b];
-                let child = match child {
-                    Some(c) => c as usize,
-                    None => {
-                        let idx = self.next.len() as u32;
-                        self.next.push([None; 256]);
-                        self.count.push(0);
-                        self.next[node][b] = Some(idx);
-                        idx as usize
-                    }
-                };
-                self.count[child] += 1;
-                node = child;
-            }
-        }
-    }
-
-    /// collect all substrings with their counts. `path` is a reusable buffer
-    /// passed through recursion; results are deterministic because children
-    /// are visited in ascending byte order.
-    fn collect(&self, out: &mut Vec<(Vec<u8>, u64)>, path: &mut Vec<u8>, node: usize) {
-        if self.count[node] > 0 {
-            out.push((path.clone(), self.count[node]));
-        }
-        for b in 0..256u16 {
-            if let Some(child) = self.next[node][b as usize] {
-                path.push(b as u8);
-                self.collect(out, path, child as usize);
-                path.pop();
-            }
-        }
-    }
-}
-
-impl SymbolTable {
-    /// greedily build a static table from a frequency pass over `corpus`.
-    /// deterministic: candidate substrings are counted and selected by
-    /// (count desc, bytes asc), so two builds match exactly.
-    fn build(corpus: &[u8]) -> Self {
-        // count every 1..=MAX_SYMBOL_LEN substring occurrence with a byte trie
-        // to avoid per-substring allocations. the collector visits children in
-        // ascending byte order, so the resulting ranking is deterministic.
-        let mut freq = FreqTrie::new();
-        freq.count_substrings(corpus, MAX_SYMBOL_LEN);
-        let mut ranked: Vec<(Vec<u8>, u64)> = Vec::new();
-        let mut path: Vec<u8> = Vec::new();
-        freq.collect(&mut ranked, &mut path, 0);
-
-        // rank by estimated saved bytes desc, then bytes asc for determinism.
-        ranked.sort_by(|a, b| {
-            let gain_a = (a.0.len() as u64 - 1) * a.1;
-            let gain_b = (b.0.len() as u64 - 1) * b.1;
-            gain_b.cmp(&gain_a).then_with(|| a.0.cmp(&b.0))
-        });
-        // always include every single byte that appears, so no input ever
-        // needs more escapes than necessary; then fill the rest with the
-        // highest-gain multi-byte symbols up to N_CODES.
-        let mut symbols: Vec<Vec<u8>> = Vec::with_capacity(N_CODES);
-        let mut seen: std::collections::BTreeSet<Vec<u8>> = std::collections::BTreeSet::new();
-        for (sym, _) in &ranked {
-            if symbols.len() >= N_CODES {
-                break;
-            }
-            if seen.insert(sym.clone()) {
-                symbols.push(sym.clone());
-            }
-        }
-        let mut trie = Trie::new();
-        for (code, sym) in symbols.iter().enumerate() {
-            trie.insert(sym, code as u8);
-        }
-        Self { symbols, trie }
-    }
-
-    /// longest table symbol matching `input` at its start, returning
-    /// `(code, length)`, or `None` if no multi/single-byte symbol matches
-    /// (then the caller escapes the single raw byte).
-    fn longest_match(&self, input: &[u8]) -> Option<(u8, usize)> {
-        self.trie.longest_match(input)
-    }
-}
-
-/// serialize the table: u16 count (LE) + per symbol a u8 length + bytes.
-fn serialize_table(table: &SymbolTable) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.extend_from_slice(&(table.symbols.len() as u16).to_le_bytes());
-    for sym in &table.symbols {
-        out.push(sym.len() as u8);
-        out.extend_from_slice(sym);
-    }
-    out
-}
-
-/// parse a serialized table, bounds-checked. returns the table and the byte
-/// length consumed so the caller can locate what follows it.
-fn parse_table(bytes: &[u8]) -> crate::Result<(Vec<Vec<u8>>, usize)> {
-    if bytes.len() < 2 {
-        return Err(malformed("fsst: truncated table count"));
-    }
-    let count = u16::from_le_bytes([bytes[0], bytes[1]]) as usize;
-    if count > N_CODES {
-        return Err(malformed("fsst: table count exceeds 255"));
-    }
-    let mut pos = 2;
-    let mut symbols = Vec::with_capacity(count);
-    for _ in 0..count {
-        let len = *bytes
-            .get(pos)
-            .ok_or_else(|| malformed("fsst: truncated symbol len"))? as usize;
-        if len == 0 || len > MAX_SYMBOL_LEN {
-            return Err(malformed("fsst: symbol len out of range"));
-        }
-        pos += 1;
-        let sym = bytes
-            .get(pos..pos + len)
-            .ok_or_else(|| malformed("fsst: truncated symbol bytes"))?
-            .to_vec();
-        pos += len;
-        symbols.push(sym);
-    }
-    Ok((symbols, pos))
-}
 
 /// encode one string with `table`: greedy longest-match, escape any byte no
 /// symbol covers. lossless for arbitrary bytes.
@@ -319,8 +108,6 @@ pub fn encode(texts: &[String]) -> crate::Result<Vec<u8>> {
 /// `sections::encode_chunks_canonical`, so `content_hash` is preserved.
 pub fn decode(bytes: &[u8]) -> crate::Result<Vec<u8>> {
     let (count, offsets, framed) = parse_v3(bytes)?;
-    // framed = u32 table_len + serialized table + streams region. the offset
-    // table indexes into the streams region (past the table blob).
     if framed.len() < 4 {
         return Err(malformed("fsst: truncated region header"));
     }
